1_foundation/requirements.txt
openai
pypdf
requests
gradio
python-dotenv